Move web-specific NetworkInterceptor into web
